Fix use of printf like functions

+++ mapgdal.c
@@ -468,7 +468,7 @@ int msSaveImageGDAL( mapObj *map, imageObj *image, char *filename )
   if (papszOptions == NULL) {
     msReleaseLock( TLOCK_GDAL );
     msSetError( MS_MEMERR, "Out of memory allocating %u bytes.\n", "msSaveImageGDAL()",
-                sizeof(char *)*(format->numformatoptions+1));
+                (unsigned int)(sizeof(char *)*(format->numformatoptions+1)));
     return MS_FAILURE;
   }
